+++ src/ThemeManager.ts
@@ -1,9 +1,10 @@
 import * as vscode from "vscode";
-import { DokiTheme } from "./DokiTheme";
-import { installSticker, removeStickers, InstallStatus } from "./StickerService";
-import { VSCodeGlobals } from "./VSCodeGlobals";
-import { StatusBarComponent } from "./StatusBar";
-import { showStickerInstallationSupportWindow, showStickerRemovalSupportWindow } from "./SupportService";
+import {DokiTheme} from "./DokiTheme";
+import {InstallStatus, installSticker, removeStickers} from "./StickerService";
+import {VSCodeGlobals} from "./VSCodeGlobals";
+import {StatusBarComponent} from "./StatusBar";
+import {showStickerInstallationSupportWindow, showStickerRemovalSupportWindow} from "./SupportService";
+import DokiThemeDefinitions from "./DokiThemeDefinitions";
 
 export const ACTIVE_THEME = 'doki.theme.active';
 
@@ -27,18 +28,17 @@ async function attemptToInstall(
 
     if (result && result.title === stickerInstall) {
       context.globalState.update(FIRST_TIME_STICKER_INSTALL, true);
-      const installStatus = performStickerInstall(dokiTheme);
-      return installStatus;
+      return performStickerInstall(dokiTheme, context);
     } else {
       return InstallStatus.NOT_INSTALLED;
     }
   } else {
-    return performStickerInstall(dokiTheme);
+    return performStickerInstall(dokiTheme, context);
   }
 }
 
-function performStickerInstall(dokiTheme: DokiTheme) {
-  const installResult = installSticker(dokiTheme);
+function performStickerInstall(dokiTheme: DokiTheme, context: vscode.ExtensionContext) {
+  const installResult = installSticker(dokiTheme, context);
   return installResult ? InstallStatus.INSTALLED :
     InstallStatus.FAILURE;
 }
@@ -49,7 +49,7 @@ export function activateTheme(
 ) {
   attemptToInstall(dokiTheme, context).then(didInstall => {
     if (didInstall === InstallStatus.INSTALLED) {
-      VSCodeGlobals.globalState.update(ACTIVE_THEME, dokiTheme.displayName);
+      VSCodeGlobals.globalState.update(ACTIVE_THEME, dokiTheme.id);
       StatusBarComponent.setText(dokiTheme.displayName);
       vscode.window.showInformationMessage(`${dokiTheme.name} installed!\n Please restart your IDE`);
     } else if (didInstall === InstallStatus.FAILURE) {
@@ -72,3 +72,14 @@ export function uninstallImages(
     vscode.window.showErrorMessage(`Unable to remove stickers/background, please see active tab for more information.`);
   }
 }
+
+export const getCurrentTheme = (): DokiTheme => {
+  const currentThemeId = VSCodeGlobals.globalState.get(ACTIVE_THEME);
+  const dokiThemeDefinition = DokiThemeDefinitions.find(
+    dokiDefinition => 
+    dokiDefinition.themeDefinition.information.id === currentThemeId ||
+    // todo: remove this after deploy.
+    dokiDefinition.themeDefinition.information.displayName == currentThemeId
+  ) || DokiThemeDefinitions[0];
+  return new DokiTheme(dokiThemeDefinition.themeDefinition);
+};