+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2.0.2 [Non-Functional]
+
+- Changed how the plugin gets stickers. 
+    - I am now able to update the stickers without you having to download a new version.
+
+
 ## 2.0.1 [Code Server Support]
 
 - Stickers/background can be installed on VSCode running on Code Server.